Add basic functionality
Duke class: add printWithBorders() method

This method takes a string as input and prints it within borders before
and after the string.

Duke class: add while-loop within main() to provide basic functionality

The while-loop receives inputs from the user and echos it back.
If the string "bye" is given, the program terminates with a goodbye message.