modified Duke.java to provide functionality for Level-2
Implemented a String array (named `records`) to contain user entries.
Since it can be assumed that there will be at most 100 entries, a fixed size
array initialised to 100 was implemented.

Modified the while-loop in Duke.main() to add entries to `records`.

Modified the while-loop in Duke.main() to list out entries in `records`.