Add support for different types of Tasks
There are different types of tasks and their properties
should be taken into account.

Created three subtypes of Task. Changed Task to a mutable
class and allow toggling of the completion status.
Parsing of the input strings by split() is implemented.

The immutable state of the Task class is undesirable
for inheritance. Parsing method is made to fit the
command-description style of input.