Abstract chat functions to a separate class
The Duke class is messy and not written in an object-oriented style.

Abstracted all chatting functions to new class so that the Duke
class only does the logo printing and input looping.

The main class is made cleaner and the chatting functions are
rewritten in OOP style to clear the code and enhance readability.