Implement shell script testing
Shell script testing is used to faster debug the code.

Created test case and fixed minor bugs in the output in order to
decrease issues in the shell script, namely, the spaces at the end
of output lines are deleted.

Spaces at the end of a line cause comparability issues, and causes
the shell script to fail test cases even if outputs are the same.