Increase exception handling cases
The code should have more flexibility for the acceptance of exceptions.

Various try-catch blocks are added to catch different exception types
and to give different responses according to.

The use of Java inbuilt exceptions are used directly rather than self-
defined ones are due to the inbuilt exceptions being sufficient for
the current exceptions that may be caught.