Convert commands to enum types
The commands in the Blarb class is implemented through a long
if-else if-else block, while switch-case blocks would allow cleaner
code.

The enum type for commands are created, and the long if-else if-else
block in Blarb.java is replaced with a switch-case.

The enum type is used in order to allow for the switch-case block to
be used as it only accepted single variable types.