Implement level-4 of Duke
New features in level-4 introduce types of tasks that have common
behaviours but are still functionally diferent.

The Task class is not suited to deal with these behaviours in an
efficient manner without code duplication or messy if-else checks.

We will turn Task into an abstract class and create classes Event, Todo
and Deadline which will be concrete implementations of the Task class.

Inheritance presents a clean implementation of the needed use case and
allows for common behaviours to remain in a parent class, and the
abstract class acts as a template for all tasks while allowing unique
behaviours to be defined by the implemented class itself.