Implement level-5 of Duke
Duke does not have any error handling.

Error handling needs to be added in order to deal with user input that
does not conform to the required spec.

Adding try-catch blocks allows the chatbot to deal with incorrect user
input through the use of exceptions. A custom DukeExceptions allows us
to take care of exceptions which are within the scope of what we predict
might go wrong.

Exceptions allow us to have a standardised way of dealing with Duke
related exceptions, as opposed to if-else checks which print errors.