Add store function
The user should be able to call the same task list across multiple
iterations of program execution.

file() and refile() functions are added to store the current task list
into an output file. The file data is also transferred into the task
list upon initialization of the program.

The program has the flexibility to either append to the storage file or
completely rewrite it, hence the two functions of file() and refile().

The main class is renamed as Main.java for the removal of Duke
references to the program.