Reorganise code segments and rename program
Many methods for handling input and tasks are bundled together in the chatbot, which retains its legacy name Duke from the start.

However, having all the methods together in one place makes it difficult to find a method to debug. Also, keeping the original name is confusing.

Let's,
* separate the methods into different management classes
* rename the main program class to SurrealChat

Separating the methods into different management classes based on roles is preferable to keeping them together because it would be easier to find a method to debug. The SurrealChat name also eliminates confusion with regards to program name.