Add optional constructor and conversion to save info string for each Task class
Why the change?
Indicating a task is completed needs to be easier as the previous
method required calling markAsDone() every time after the task was
created.

As Duke only view those tasks as Task, it has no way of accessing
the additional parameters of a task like date time and date time
range.

What is being done?
Optional constructor with a isDone is added to directly declare
whether the task is done.

Declaring a toSaveInfoString() to the base Task class and have the
other task classes to override, adding on their own parameters.

Why is it done that way?
Instantiating a completed task can be done in 1 line, so when
loading completed tasks, Duke can instantiate such task without
extra lines of code.

Preserve polymorphism among the task classes. If there is new task
type to add, it can just override the toSaveInfoString() easily.