Add save/load functionality for tasks
SurrealChat does not keep track of tasks between each run.

This causes time wastage in initialising tasks again.

Let's give a save/load functionality for tasks to be saved into a file and loaded from that file in the next run.

Using a file also allows one to keep track of tasks outside of the program.