Split dateTimeRange field of Event into Start and End fields
Why the change?
Having 1 field to represent a date time range require us to
establish a specific format to follow which may not be intuitive
to the user.

What is being done?
Replace dateTimeRange field and start and end fields whereby
the Event command syntax is now "event XXX /start XXX /end XXX".
Related classes and functions are updated to reflect the change.

Why is it done that way?
Since the date time range will be represented with 2 fields, instead
of /at, using /start and /end comes more naturally to both the
developers and the users.