Add feature "Save data to hard disk, load data from hard disk"
Add a "FileHandler" class with two functions (read and update).

The file will be updated only when the program terminates instead
of appending tasks to the file when user is using. It might
not be that efficient, but the number of tasks from the user
will not be very large, hence the performance will be okay.

However, the file path specified is not an OS-independent way.
Hence, changes need to be made to let Duke be OS-independent.