Replace manual printing of welcome, goodbye and error messages with Ui class
Why the change?
Extract out the printing aspect of Duke into a seperate Ui class, in
an effort to keep the Duke class managable.

What is being done?
Manual printings of welcome, goodbye and error message are replaced
with functions from Ui class.

Why is it done that way?
By seperating the printing aspect to another class, any printing
issues will be isolated into that one class, allowing the debugging
to be easier.