Compartmentalise into classes
File operations as well as output and error print statements are scattered around the code.

However, these make it harder to modify print output and file operations as one would have to jump from class to class.

Let's,
* add a UserOutput class to handle print-related code
* add a FileManagement class to handle file operations
* add a TaskParser class to act as an intermediary for processing of tasks between IO side and TaskManagement side
* add a Pair class to store a command and an associated generic type item for further processing in UserOutput
* rename InputManagement to UserInput
* remove text-ui-test cases temporarily

Compartmentalising tasks makes it easier to search for the relevant code segment to modify and debug. Renaming InputManagement to UserInput also emphasises on the methods being used for user commands. In addition, removing the test cases for now allows for re-working of test cases without worrying about main code and vice versa.