Extract conversion of text to executing commands as Parser class and update relevant classes to reflect the changes
Why the change?
A large portion of Duke's main function is to match input strings to
a command which can be extracted to another class, isolating the
matching process away from the Duke class.

What is being done?
The match process will be extracted into a Parser class that reads
an input string, verify the string and return a corresponding
function to be executed.

Why is it done that way?
This keeps the matching process to be isolated into 1 location, the
Parser class, allowing the debug process to be much easier.