Refactor Duke
Why the change?
It is impossible to run multiple instance of Duke as all the running
logic is in the static main() function, not an instance function.

What is being done?
All logic required for the bot (ie. setup and running) are allocated
to its constructor and run().

Why is it done that way?
It will be possible to run multiple instances of the bot if
required.