Package classes into blarb package
Software engineering projects should be appropriately packaged to ensure
that implementation details are not seen by the client.

All classes except Main are wrapped into the blarb package. Access
modifiers for packaged classes, except Blarb are set to
package-private to hide details and as the Main class only requires
the Blarb class to execute.

Most classes are set to package-private for proper encapsulation.