Add find task capability
Meme Man displays all tasks to the user.

This is impractical when the user only wishes to find certain tasks.

Let's add a find capability to SurrealChat that can take a keyword and find tasks with that keyword, regardless of upper or lower case.

This is preferable to reading through the entire list of tasks.