Change date time format for saving Deadline and Event
Why the change?
Previous format removes the year completely which will be required
when we load a save file.

What is being done?
Format is change to included year as well. Since the day of the week
in english word is not required, it will be remove from the format.

Why is it done that way?
If future features ever need to search tasks by year, this allowed
tasks loaded from the hard disk to be searchable by that method too.