Add exception handling for incorrect date format
The Deadline and Event tasks take in a date input in the form YYYY-MM-DD. If the input is incorrect, it would throw an uncaught exception.

This is not ideal as the program would terminate prematurely.

Let's,
* abstract out the code snippet used for parsing date in the methods for generating Deadline and Event tasks
* add exception handling for the incorrect date, along with a custom error message

This is preferable as code can continue to run even after the incorrect date format is keyed in and the message generated shares the same theme as other custom exception messages by SurrealChat. In addition, abstracting out the code makes changing the code easier if the date format is changed in the future.