Add ability to save and reload tasks to /from file
Modify Duke, Task, Todo, Event and Deadline to save tasks to file and reload on program startup.

Previously, tasks were lost at the end of the program.

Tasks are now saved at the end of the program ("Bye" command) and
reloaded at the start of the program. To facilitate this, saveTasks and
loadTasks method were implemented in Duke class to save and load tasks. Task
(and sub-classes) also modified to now provide a storageEntry method to
output the string formatted entry to be written to file.