Update Vergil.java and text-ui-test/EXPECTED.TXT
All of the command processing code is written in the main() method, making it difficult to read, test, and modify code behaviour. Additionally, some error messages are ambiguous and need more clarification.

Let's split the code into multiple methods that process different types of commands separately, and modify the error messages to become more specific.

Writing a standalone method for each type of commands makes the process of understanding and testing a certain behaviour of the code easier, since there would be no need to read through and modify some or all of the other behaviours. In addition, specifing the issues in the error messages reduces the time wasted to understand the typing mistakes in the commands.