Add save to disk functionality
Data about tasks are stored in memory and will not preserve to the next
section.

Let's write existing data about user's todo into a file under
current working directory and into /data/database.txt

This is to ensure on starting up the program again, data can be read
from the file and display to users.