Extract date from Event description
For Deadline and Event classes, we need a way to parse and store the dates associated with these two classes.

Static methods in Parser allows for parsing based on the date. So far the only recognisable format is  yyyy-MM-DD. F
uture iterations could have more date formats.

This would facilitate future features which can categorise events based on their dates.