Handle file or directory is missing
There might be a problem when the program is run on other computers where they do not have the directories or files needed by duke. In this case we need to handle it by creating the files or directories for them.

Additionally added Duke constructor where the missing file or directory case will be handled.