Parse date and time for user input
Date and time are treated as pure string, which cannot be parsed and
compared later on.

Let's parse the input string and store date and time information as
LocalDate and LocalTime object. Also formatting is applied when printing
the information.

This helps users to see the date and time information easily.