Implement download data but not save data
DukeBot CLI does not read or store data locally, persisting data for
only as long as it the program is running.

By not saving data, the DukeBot cannot be used across different
sessions.

Add a way to read data in from a text file.

This commit is incomplete in that it does not include a way to write
data to a text file.

The ability to write to a file requires us to store the Tasks completion
status, something that needs to be further planned out to utilise
existing code.