Organise into packages and implement TaskList class wrapper
Organise the classes into packages as there are too many classes and some needs to be logically grouped together. The new TaskList class also acts as a wrapper over TaskList and seperates it out from the main.

Organising into packages is neater and hence we can find files fast. It also helps with maintainability. The TaskList class adheres more strongly to OOP design.