Persist the task list in memory and load from memory on start-up
DukeBot runs with the Task List in RAM and does not persist it beyond
each session.

Temporary memory prevents the task list from being useful to a user
beyond a single session, requiring the user to keep DukeBot running.

Updates DukeBot to store the contents of its taskList to a file in
memory, allowing information to be persisted across multiple sessions.

There are some remaining To Do's involving refactoring the IO code into
separate functions and updating the Java Docs.