Add Save functionality
Duke is unable to save the task list on the local computer itself.

The software might crash and cause data to be lost. Thus a backup
is needed.

Save the task list on the local computer serves as a backup in case
data is lost.

Saving the task list on the local computer is convenient and
also reliable.