Add edit task description functionality
Tasks are added with fixed descriptions.

Editing task descriptions is unfeasible since one has to delete the task and add a new task with the description.

Let's,
* add a functionality to edit task descriptions without deleting the task.
* make tasks immutable.
* add new JUnit tests to test editing of tasks and marking tasks as done or undone.

This is preferable as task descriptions can be conveniently edited. In addition, immutability helps protect the internal variables of the Task class from inadvertent editing.