Seperate out Parser Logic from Main.
Adhering to true OOP principles, restructure the code so that it is better organised . The Parser class handles most of the logic of reading input String and figuring out what Command to return.
Command.execute() carries out the calling of the UI and the storage. There is more work to be done to seperate out the Ui Logic from the Command class.

** TODO ** Seperate out the printing to the ui class. Also could seperate out Date Parsing logic from the Command Parser class . For future iterations/