Add J-Unit Tests for Deadline.
Add JUnit tests to test Deadline. Check two methods for correctness, the toString method,
and the method for generating the saved String format used for storing the task in the file.

There could be instances where there is a typo that results in an incorrect String format.
This might lead to bugs when the tasks are being stored in another .txt file and fetched the next
time the app is open.

J-Unit tests help to automate the testing.

Let's add unit test to check for toString method correctness.
a unit test to check the getSavedString format correctness
and a unit test to check if Deadline throws Exception when
no date is able to be parsed from its by String.