Add  JavaDoc Comments and extract duplicate code
There is duplicate code in the toString method of various subclasses of Task. The common code can be extracted out under the Task class,
hence shortening the code. JavaDoc comments are needed to document what each class does and to make it easier to read.

Let's put JavaDoc comments for at least half of the classes, with the rest to be filled in at later dates.
Let's extract out the this.getLetterCode() function call  to the Task class method.