Reorganised code to follow OOP principles
Created 4 classes to deal with general usage:
Ui - to deal with interactions with the user (printing messages and reading inputs)
Storage - to deal with loading/writing tasks from/to file in hard drive
Parser - to deal with different commands input from user
TaskList - to deal with organising the tasks (add/delete/markDone etc)

Created another 6 classes with a superclass Command and 5 subclasses
AddCommand - to deal with adding different tasks
DeleteCommand - to deal with deleting a certain task
DoneCommand - to deal with marking a certain task as done
ExitCommand - to deal with exiting the program
ListCommand - to deal with displaying the list of tasks at hand

Some exceptions are not thrown/handled properly yet. There are some cases where the program will be unable to deal with.
Example: program does not deal with deletion/marking of a task that does not exist