Extract out field isExit from Command class
Previously, the command class execute method would, after execution, return a boolean indicating
if the program should terminate or not. This is harder to read as one would have to look
for the meaning of the returned boolean value from execute.

 iI the boolean variable were to be separated out as a  field within the command c it would be easier.

Let's modify execute to be a void method, while the Command
class contains the attribute isExit  and method shouldExitt(),
which should inform the user if the command is telling the program to exit.