Rework code for commands and output
Code for handling user commands and file loading is mainly handled by TaskParser class. This also generates pairs of an internal tag and a generic second item of any type. The done and undone commands are also separate from each other.

This centralisation of code is confusing to look at. Having generics in UserOutput also results in compile warnings. A separate Done and Undone command not only duplicates code but also causes confusion should an undo/redo function be added in future.

Let's,
* reorganise code related to user commands into Command and EasterEgg classes and their respective subclasses.
* package the base .java files into actual packages based on package statements.
* generate the string outputs within the respective classes, rather than at the UserOutput class.
* move code for handling file operations from TaskParser class to TaskManagement class.
* consolidate done and undone command into done command to toggle a task status between done and undone.

This is preferable as editing code for an individual command can now be easily done at that command class. In addition, there is no need for the Pair class to carry unprocessed output for further processing at UserOutput and for the TaskParser class to handle user commands and file operations. Both the Pair class and TaskParser class can be removed. There is no more confusion between Undone task command marking a task as undone and a potential undo command for rolling back a certain change.