Connect the chatbot to the GUI
Handle the user input, and give the corresponding duke reply.
Create two new packages: duke.gui, duke.task. Because of the
packaging, the constructor of each subclass of tasks should become
public. Move most of the things into ChatBot class, and Duke class
become a class only be used for CLI.

Modify the return values of functions in TaskList class from void to String,
so that it can be used in handleUserInput function in MainWindow class.
However, the JavaDoc remains the same for now. Hence, a lot of
JavaDoc needs to be changed.