Add help functionality
The README.md user guide is the only place to find guidance on how to use commands in SurrealChat.

This is inconvenient for people who need to use SurrealChat on the fly.

Let's add a help command that allows the user to read a brief description on how to use a command on the spot without having to pause to look up the user guide.

This is preferable as the user can quickly see the brief but important details on command usage and only need to refer to the user guide for examples and exception messages.