Refactor the code into OOP classes for UI, Parser, etc.
Each class in DukeBot handles its own IO and Parsing causing code to be
harder to maintain in the long run.

Spreading the IO and Parsing across many different classes makes DukeBot more
difficult to maintain in the long run.

Includes dedicated classes to handle these responsibilities.

All classes are now self-contained and hence any future changes to the
codebase can be targeted at whichever class is solely responsible for
the feature.

The Exception handling functionality, including the use of try-catch
blocks is not very elegant.