Include JUnit tests for Storage and Parser
DBot does not have any JUnit unit tests for any classes or scenarios.

The lack of unit testing allows for potential regressions in any future
edits to the code base.

Includes JUnit tests for Storage initialisation on invalid file paths
and Parser.parse on a valid Todo command.

These tests are not comprehensive and more test cases are needed to
prevent regressions.