Make search more intuitive
Search is currently only a wrapper on contains()

This is bad because this breaks intuition (At least my intuition) about
how searching should be done. For example: "in" should not match "Beating Minecraft"
and "minecraft" should match it, but both of those cases are unfulfilled.

Let's:
* Attempt to check description on a per word basis from the start of the word
** Allow this to be overriden by the presence of multiple spaces
* Do lowercase matching for everything by default
** Allow capital letters in the search to override this.

The reason we allow exceptions is because we sometimes want more specific searching rules.