Include JavaDoc for all of DBot's Classes and Methods
DBot's implementation does not have many JavaDoc comments to explain
what each class or method is responsible for, how they are to be used,
or what the arguments they take represent.

This lack of JavaDoc comments can make it difficult for future
developers to use the DBot package.

Let's include JavaDoc comments for every Class or Method to assist
future developers.

JavaDoc comments can also be used by IDE's to quickly supply developers
with the information they need when dealing with a class or method.