Refactor code to make the code have better code quality
The code follows basic java coding standards, but some of the code
is written in a way that does not follow best practices of code quality.

Code of poor code quality can propogate the messy effects to the future.

Let's refactor the code to make it have better code quality.