Prevent the label truncating long text
When the total height of all the DialogueBoxes exceed the display
height, the height of every DialogueBox is adjusted to compensate for
it, until a minimum of 2 lines are shown. This is not desired since it
will mean that text is truncated which does not make sense for a chat
dialogue.

This is resolved by setting the minHeight of the DialogueBox to
USE_PREF_HEIGHT, and therefore the DialogueBox will maintain the correct
height.