Expand edit functionality
Edit allows for the description of the task to be changed without deleting the task.

This is still inconvenient for changing date time and priority level.

Let's expand edit such that it can also change date time and/or priority level.

This allows for all details regarding a task to be edited without having to delete the task.