Create new class Storage Parser
Extract method refactoring done on Parser to extract out methods used for parsing the
encoded task string from database.

This is so that it is neater and Parser class will not be responsible for doing different unrelated things.
Parser will seperate out to CommandParser and StorageParser. Command Parser will just focus on
parsing Commands.