Reset the dialog box's maximum height
The dialog box maximum height is the default value.
When too much text comes into the dialog box, it is possible for the dialog
box not to contain all information.

The dialog should be able to auto-adjust the maximum height
with the text it is displaying.

Let us update the dialog box so that the desired functionality is achieved.