Let bot change background color when the user input is incorrect to remind user
The bot does not change background color.

User may want to be reminded if their input is not in the correct format.

Let us change the code so that the bot can change background color if the
user input is incorrect. Right now we do not care how the checking if input
is correct, but we focus on how to change color. The checking will be done in
the next commit.